requirements.txt
pytest
requests
allure-pytest
#список всех библиотек, которые нужны для проекта